Various small improvements and fixes
Add indicator for external links.
Bug: Linked images also have this indicator.

Remove empty „send“ button in the search form.

Add heading for the search field on the archive
page for better markup and accessibility.

Optimize appearance of the category list on the
archive page and fix an HTML markup issue.

Add rel prev/next attributes to the pagination for
better accessibility.

Add blog name to the footer.

Add RSS link to the footer.

Add link to the CC license in the footer.

Reverse backend editor background color to white
for standard behavior.